# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Accountant Practice Backoffice API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Client {
  id: number;
  code: string;
  name: string;
  type: string;
  /** @nullable */
  addressLine1?: string | null;
  /** @nullable */
  addressLine2?: string | null;
  /** @nullable */
  town?: string | null;
  /** @nullable */
  county?: string | null;
  /** @nullable */
  postcode?: string | null;
  /** @nullable */
  contactNumber?: string | null;
  /** @nullable */
  email?: string | null;
  createdAt?: string;
}

export interface ClientInput {
  code: string;
  name: string;
  type: string;
  /** @nullable */
  addressLine1?: string | null;
  /** @nullable */
  addressLine2?: string | null;
  /** @nullable */
  town?: string | null;
  /** @nullable */
  county?: string | null;
  /** @nullable */
  postcode?: string | null;
  /** @nullable */
  contactNumber?: string | null;
  /** @nullable */
  email?: string | null;
}

export interface ClientUpdate {
  code?: string;
  name?: string;
  type?: string;
  /** @nullable */
  addressLine1?: string | null;
  /** @nullable */
  addressLine2?: string | null;
  /** @nullable */
  town?: string | null;
  /** @nullable */
  county?: string | null;
  /** @nullable */
  postcode?: string | null;
  /** @nullable */
  contactNumber?: string | null;
  /** @nullable */
  email?: string | null;
}

export interface ClientList {
  clients: Client[];
  total: number;
  page: number;
  limit: number;
}

export interface Task {
  id: number;
  /** @nullable */
  clientId: number | null;
  clientName?: string;
  /** @nullable */
  activityType?: string | null;
  /** @nullable */
  assignedTo?: string | null;
  taskName: string;
  taskStatus: string;
  /** @nullable */
  dueDate?: string | null;
  isOverdue?: boolean;
}

export interface FinancialInfo {
  id: number;
  clientId: number;
  /** @nullable */
  clientCode?: string | null;
  /** @nullable */
  clientType?: string | null;
  /** @nullable */
  turnover?: number | null;
  /** @nullable */
  profitBeforeTax?: number | null;
  /** @nullable */
  totalIncome?: number | null;
  /** @nullable */
  totalSelfEmploymentIncome?: number | null;
  /** @nullable */
  totalProfitFromSelfEmployments?: number | null;
  /** @nullable */
  avgEmployees?: number | null;
}

export interface TaxReference {
  id: number;
  clientId: number;
  /** @nullable */
  clientName?: string | null;
  /** @nullable */
  utr?: string | null;
  /** @nullable */
  payeRef?: string | null;
  /** @nullable */
  payeAccountsOfficeRef?: string | null;
  /** @nullable */
  vatRegNo?: string | null;
  /** @nullable */
  vatRegDate?: string | null;
  /** @nullable */
  taxOffice?: string | null;
  /** @nullable */
  niNumber?: string | null;
  /** @nullable */
  companyRegNo?: string | null;
  /** @nullable */
  dateOfIncorporation?: string | null;
  /** @nullable */
  amlStatus?: string | null;
  /** @nullable */
  latestAccountsStatus?: string | null;
  /** @nullable */
  engagementStatus?: string | null;
}

export interface TaxReturn {
  id: number;
  /** @nullable */
  clientId: number | null;
  /** @nullable */
  clientCode?: string | null;
  clientName?: string;
  /** @nullable */
  clientType?: string | null;
  /** @nullable */
  taxReturnStatus?: string | null;
}

export interface ClientDetail {
  client: Client;
  tasks?: Task[];
  financialInfo?: FinancialInfo | null;
  taxReference?: TaxReference | null;
  taxReturn?: TaxReturn | null;
}

export interface TaskInput {
  /** @nullable */
  clientId?: number | null;
  /** @nullable */
  clientName?: string | null;
  /** @nullable */
  activityType?: string | null;
  /** @nullable */
  assignedTo?: string | null;
  taskName: string;
  taskStatus: string;
  /** @nullable */
  dueDate?: string | null;
}

export interface TaskUpdate {
  /** @nullable */
  clientId?: number | null;
  /** @nullable */
  clientName?: string | null;
  /** @nullable */
  activityType?: string | null;
  /** @nullable */
  assignedTo?: string | null;
  taskName?: string;
  taskStatus?: string;
  /** @nullable */
  dueDate?: string | null;
}

export interface TaskList {
  tasks: Task[];
  total: number;
  page: number;
  limit: number;
}

export interface FinancialInfoUpdate {
  /** @nullable */
  turnover?: number | null;
  /** @nullable */
  profitBeforeTax?: number | null;
  /** @nullable */
  totalIncome?: number | null;
  /** @nullable */
  totalSelfEmploymentIncome?: number | null;
  /** @nullable */
  totalProfitFromSelfEmployments?: number | null;
  /** @nullable */
  avgEmployees?: number | null;
}

export interface TaxReferenceUpdate {
  /** @nullable */
  utr?: string | null;
  /** @nullable */
  payeRef?: string | null;
  /** @nullable */
  payeAccountsOfficeRef?: string | null;
  /** @nullable */
  vatRegNo?: string | null;
  /** @nullable */
  vatRegDate?: string | null;
  /** @nullable */
  taxOffice?: string | null;
  /** @nullable */
  niNumber?: string | null;
  /** @nullable */
  companyRegNo?: string | null;
  /** @nullable */
  dateOfIncorporation?: string | null;
  /** @nullable */
  amlStatus?: string | null;
  /** @nullable */
  latestAccountsStatus?: string | null;
  /** @nullable */
  engagementStatus?: string | null;
}

export interface TaxReturnUpdate {
  /** @nullable */
  taxReturnStatus?: string | null;
}

export interface TaxReturnList {
  taxReturns: TaxReturn[];
  total: number;
  page: number;
  limit: number;
}

export interface StatusCount {
  label: string;
  count: number;
}

export interface DashboardSummary {
  totalClients: number;
  overdueTasksCount: number;
  activeClientsCount: number;
  totalTasks?: number;
  tasksByStatus: StatusCount[];
  clientsByType: StatusCount[];
  taxReturnsByStatus: StatusCount[];
  /** @nullable */
  totalTurnover?: number | null;
  vatRegisteredCount?: number;
}

export type ListClientsParams = {
  search?: string;
  type?: string;
  engagementStatus?: string;
  page?: number;
  limit?: number;
};

export type ListTasksParams = {
  clientId?: number;
  status?: string;
  activityType?: string;
  assignedTo?: string;
  overdue?: boolean;
  page?: number;
  limit?: number;
};

export type ListFinancialInfoParams = {
  clientId?: number;
  clientType?: string;
};

export type ListTaxReferencesParams = {
  clientId?: number;
  engagementStatus?: string;
  amlStatus?: string;
};

export type ListTaxReturnsParams = {
  clientId?: number;
  status?: string;
  clientType?: string;
  search?: string;
  page?: number;
  limit?: number;
};

export type GetDashboardOverdueTasksParams = {
  limit?: number;
};

export type GetDashboardRecentClientsParams = {
  limit?: number;
};