# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Accountant Practice Backoffice API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tax-references:
    get:
      operationId: listTaxReferences
      tags: [tax-references]
      summary: List tax references
      parameters:
        - name: clientId
          in: query
          schema:
            type: integer
        - name: engagementStatus
          in: query
          schema:
            type: string
        - name: amlStatus
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of tax references
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TaxReference"
  /tax-references/{clientId}:
    get:
      operationId: getTaxReference
      tags: [tax-references]
      summary: Get tax reference for a client
      parameters:
        - name: clientId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Tax reference
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaxReference"
        "404":
          description: Not found
    patch:
      operationId: updateTaxReference
      tags: [tax-references]
      summary: Update tax reference for a client
      parameters:
        - name: clientId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaxReferenceUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaxReference"
  /tax-returns:
    get:
      operationId: listTaxReturns
      tags: [tax-returns]
      summary: List SA tax return statuses
      parameters:
        - name: clientId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
        - name: clientType
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of tax return records
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaxReturnList"
  /tax-returns/{id}:
    patch:
      operationId: updateTaxReturn
      tags: [tax-returns]
      summary: Update a tax return record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaxReturnUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaxReturn"
...
components:
  schemas:
...
    TaxReference:
      type: object
      required: [id, clientId]
      properties:
        id:
          type: integer
        clientId:
          type: integer
        clientName:
          type: ["string", "null"]
        utr:
          type: ["string", "null"]
        payeRef:
          type: ["string", "null"]
        payeAccountsOfficeRef:
          type: ["string", "null"]
        vatRegNo:
          type: ["string", "null"]
        vatRegDate:
          type: ["string", "null"]
        taxOffice:
          type: ["string", "null"]
        niNumber:
          type: ["string", "null"]
        companyRegNo:
          type: ["string", "null"]
        dateOfIncorporation:
          type: ["string", "null"]
        amlStatus:
          type: ["string", "null"]
        latestAccountsStatus:
          type: ["string", "null"]
        engagementStatus:
          type: ["string", "null"]
    TaxReferenceUpdate:
      type: object
      properties:
        utr:
          type: ["string", "null"]
        payeRef:
          type: ["string", "null"]
        payeAccountsOfficeRef:
          type: ["string", "null"]
        vatRegNo:
          type: ["string", "null"]
        vatRegDate:
          type: ["string", "null"]
        taxOffice:
          type: ["string", "null"]
        niNumber:
          type: ["string", "null"]
        companyRegNo:
          type: ["string", "null"]
        dateOfIncorporation:
          type: ["string", "null"]
        amlStatus:
          type: ["string", "null"]
        latestAccountsStatus:
          type: ["string", "null"]
        engagementStatus:
          type: ["string", "null"]
    TaxReturn:
      type: object
      required: [id, clientId]
      properties:
        id:
          type: integer
        clientId:
          type: ["integer", "null"]
        clientCode:
          type: ["string", "null"]
        clientName:
          type: string
        clientType:
          type: ["string", "null"]
        taxReturnStatus:
          type: ["string", "null"]
    TaxReturnUpdate:
      type: object
      properties:
        taxReturnStatus:
          type: ["string", "null"]
    TaxReturnList:
      type: object
      required: [taxReturns, total, page, limit]
      properties:
        taxReturns:
          type: array
          items:
            $ref: "#/components/schemas/TaxReturn"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
...